+++ 臺灣言語服務/Moses介面.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import Pyro4
-from django.http.response import HttpResponse, JsonResponse
+from django.http.response import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 
 
@@ -10,7 +10,7 @@ class Moses介面:
         self.服務 = Pyro4.Proxy("PYRONAME:Moses服務")
 
     def 正規化翻譯支援腔口(self, request):
-        return self.json包做回應({'腔口': self.服務.支援腔口()})
+        return JsonResponse({'腔口': self.服務.支援腔口()})
 
     @csrf_exempt
     def 正規化翻譯(self, request):
@@ -29,7 +29,7 @@ class Moses介面:
         except:
             查詢語句 = '你好嗎？我很好！'
         try:
-            return self.服務.正規化翻譯實作(查詢腔口, 查詢語句)
+            return JsonResponse(self.服務.正規化翻譯實作(查詢腔口, 查詢語句))
         except ConnectionRefusedError:
             return JsonResponse({'失敗': '服務啟動中，一分鐘後才試'})
 
@@ -50,18 +50,6 @@ class Moses介面:
         except:
             查詢語句 = '你好嗎？我很好！'
         try:
-            return self.服務.標漢字音標實作(查詢腔口, 查詢語句)
+            return JsonResponse(self.服務.標漢字音標實作(查詢腔口, 查詢語句))
         except ConnectionRefusedError:
             return JsonResponse({'失敗': '服務啟動中，一分鐘後才試'})
-
-    def json包做回應(self, json):
-        回應 = JsonResponse(json)
-        return 回應
-
-    def 音檔包做回應(self, 音檔):
-        回應 = HttpResponse()
-        回應.write(音檔)
-        回應['Content-Type'] = 'audio/wav'
-        回應['Content-Length'] = len(音檔)
-        回應['Content-Disposition'] = 'filename="taiwanese.wav"'
-        return 回應
